3_react_agent/pyproject.toml
[project]
name = "3-react-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "langchain>=0.3.27",
    "langchain-openai>=0.3.35",
    "python-dotenv>=1.1.1",
]
